chore: complete suggested changes
